backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
email-validator==2.2.0
python-multipart==0.0.12
redis==5.0.1
fastapi-limiter==0.1.6

